Add files via upload
Used getDate method in Utils class and this caused an error in the running of the code utils.getDate which did not allow the format of the date to be entered in the OrderControllerTest and OrderDAOTest. To fix this error, instead of implementing a Date method, I could have input the date as a string instead.